+++ dioxus-ui/src/components/mod.rs
@@ -1,0 +1,27 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+pub mod attendants;
+pub mod browser_compatibility;
+pub mod call_timer;
+pub mod config_error;
+pub mod device_selector;
+pub mod device_settings_modal;
+pub mod diagnostics;
+pub mod google_sign_in_button;
+pub mod host;
+pub mod host_controls;
+pub mod icons;
+pub mod login;
+pub mod meeting_ended_overlay;
+pub mod meeting_info;
+pub mod meetings_list;
+pub mod neteq_chart;
+pub mod okta_sign_in_button;
+pub mod top_bar;
+pub mod video_control_buttons;
+pub mod waiting_room;
+
+mod canvas_generator;
+mod peer_list;
+mod peer_list_item;
+mod peer_tile;