+++ videocall-transport/CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.2](https://github.com/security-union/videocall-rs/compare/videocall-transport-v0.1.1...videocall-transport-v0.1.2) - 2026-02-23
+
+### Other
+
+- updated the following local packages: videocall-types
+
+## [0.1.1](https://github.com/security-union/videocall-rs/compare/videocall-transport-v0.1.0...videocall-transport-v0.1.1) - 2026-02-19
+
+### Other
+
+- update Cargo.toml dependencies
+
 ## [0.1.0](https://github.com/security-union/videocall-rs/releases/tag/videocall-transport-v0.1.0) - 2026-02-19
 
 ### Other